+++ relay-server/src/actors/events.rs
@@ -269,22 +269,29 @@ pub struct EventProcessor {
 }
 
 impl EventProcessor {
-    #[cfg(feature = "processing")]
-    pub fn new(
-        config: Arc<Config>,
-        rate_limiter: Option<RedisRateLimiter>,
-        geoip_lookup: Option<Arc<GeoIpLookup>>,
-    ) -> Self {
+    #[inline]
+    pub fn new(config: Arc<Config>) -> Self {
         Self {
             config,
-            rate_limiter,
-            geoip_lookup,
+            #[cfg(feature = "processing")]
+            rate_limiter: None,
+            #[cfg(feature = "processing")]
+            geoip_lookup: None,
         }
     }
 
-    #[cfg(not(feature = "processing"))]
-    pub fn new(config: Arc<Config>) -> Self {
-        Self { config }
+    #[cfg(feature = "processing")]
+    #[inline]
+    pub fn with_rate_limiter(mut self, rate_limiter: Option<RedisRateLimiter>) -> Self {
+        self.rate_limiter = rate_limiter;
+        self
+    }
+
+    #[cfg(feature = "processing")]
+    #[inline]
+    pub fn with_geoip_lookup(mut self, geoip_lookup: Option<Arc<GeoIpLookup>>) -> Self {
+        self.geoip_lookup = geoip_lookup;
+        self
     }
 
     /// Validates all sessions in the envelope, if any.
@@ -294,11 +301,9 @@ impl EventProcessor {
     fn process_sessions(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
         let envelope = &mut state.envelope;
         let received = state.received_at;
-        let project_id = state.project_id;
 
         let clock_drift_processor =
             ClockDriftProcessor::new(envelope.sent_at(), received).at_least(MINIMUM_CLOCK_DRIFT);
-        let client = envelope.meta().client().map(str::to_owned);
 
         envelope.retain_items(|item| {
             if item.ty() != ItemType::Session {
@@ -312,13 +317,7 @@ impl EventProcessor {
                 Ok(session) => session,
                 Err(error) => {
                     return sentry::with_scope(
-                        |scope| {
-                            scope.set_tag("project", project_id);
-                            if let Some(ref client) = client {
-                                scope.set_tag("sdk", client);
-                            }
-                            scope.set_extra("session", String::from_utf8_lossy(&payload).into());
-                        },
+                        |s| s.set_extra("session", String::from_utf8_lossy(&payload).into()),
                         || {
                             // Skip gracefully here to allow sending other sessions.
                             log::error!("failed to store session: {}", LogError(&error));
@@ -1098,9 +1097,9 @@ impl EventProcessor {
         Ok(())
     }
 
-    fn process(
+    fn process_state(
         &self,
-        message: ProcessEnvelope,
+        mut state: ProcessEnvelopeState,
     ) -> Result<ProcessEnvelopeResponse, ProcessingError> {
         macro_rules! if_processing {
             ($if_true:block $(, $if_not:block)?) => {
@@ -1110,7 +1109,6 @@ impl EventProcessor {
             };
         }
 
-        let mut state = self.prepare_state(message)?;
         self.process_sessions(&mut state)?;
 
         if state.creates_event() {
@@ -1118,7 +1116,10 @@ impl EventProcessor {
                 self.expand_unreal(&mut state)?;
             });
 
-            self.extract_event(&mut state)?;
+            self.extract_event(&mut state).map_err(|error| {
+                log::error!("failed to extract event: {}", LogError(&error));
+                error
+            })?;
 
             if_processing!({
                 self.process_unreal(&mut state)?;
@@ -1145,6 +1146,26 @@ impl EventProcessor {
         self.scrub_attachments(&mut state)?;
 
         Ok(ProcessEnvelopeResponse::from(state))
+    }
+
+    fn process(
+        &self,
+        message: ProcessEnvelope,
+    ) -> Result<ProcessEnvelopeResponse, ProcessingError> {
+        let state = self.prepare_state(message)?;
+
+        let project_id = state.project_id;
+        let client = state.envelope.meta().client().map(str::to_owned);
+
+        sentry::with_scope(
+            |scope| {
+                scope.set_tag("project", project_id);
+                if let Some(client) = client {
+                    scope.set_tag("sdk", client);
+                }
+            },
+            || self.process_state(state),
+        )
     }
 }
 
@@ -1229,11 +1250,9 @@ impl EventManager {
 
             SyncArbiter::start(
                 thread_count,
-                clone!(config, || EventProcessor::new(
-                    config.clone(),
-                    rate_limiter.clone(),
-                    geoip_lookup.clone(),
-                )),
+                clone!(config, || EventProcessor::new(config.clone())
+                    .with_rate_limiter(rate_limiter.clone())
+                    .with_geoip_lookup(geoip_lookup.clone())),
             )
         };